Re-arrange mention match unit test
