+++ src/scripts/__tests__/utility.test.ts
@@ -4,7 +4,7 @@
  * @license GPLv2
  */
 
-import * as utility from "../utility";
+import { resolvePromisesSequentially, matchAtMention, hashString, isInstanceOfOneOf } from "../utility";
 
 describe("resolvePromisesSequentially()", () => {
     it("resolves promises in order", () => {
@@ -30,7 +30,7 @@ describe("resolvePromisesSequentially()", () => {
 
         expect.assertions(1);
 
-        return utility.resolvePromisesSequentially(functions).then(() => {
+        return resolvePromisesSequentially(functions).then(() => {
             expect(order).toEqual(expectation);
         });
     });
@@ -51,30 +51,31 @@ describe("resolvePromisesSequentially()", () => {
 
         expect.assertions(1);
 
-        return expect(utility.resolvePromisesSequentially(functions)).resolves.toEqual(expectation);
+        return expect(resolvePromisesSequentially(functions)).resolves.toEqual(expectation);
     });
 
     it("passes the value of one promise to the next", () => {
-        const func = (prev) => Number.isInteger(prev) ? Promise.resolve(prev + 1) : Promise.resolve(0);
+        const func = prev => (Number.isInteger(prev) ? Promise.resolve(prev + 1) : Promise.resolve(0));
         const functions = [func, func, func];
         const expectation = [0, 1, 2];
 
-
-        return expect(utility.resolvePromisesSequentially(functions)).resolves.toEqual(expectation);
+        return expect(resolvePromisesSequentially(functions)).resolves.toEqual(expectation);
     });
 });
 
 describe("hashString", () => {
     test("the same string always results in the same value", () => {
-        const str = "a; lksdjfl;aska;lskd fjaskl;dfj al;skdjfalsjkdfa;lksdjfl;kasdjflksaf;kbfjal;skdfbjanv;slkdfjbals;dkjfslkadfj;alsdjf;oiawjef;oiawbejvf;ioawbevf;aoiwebfjaov;wifebvl";
-        expect(utility.hashString(str)).toBe(utility.hashString(str));
+        const str =
+            "a; lksdjfl;aska;lskd fjaskl;dfj al;skdjfalsjkdfa;lksdjfl;kasdjflksaf;kbfjal;skdfbjanv;slkdfjbals;dkjfslkadfj;alsdjf;oiawjef;oiawbejvf;ioawbevf;aoiwebfjaov;wifebvl";
+        expect(hashString(str)).toBe(hashString(str));
     });
 
     test("different strings hash to different values", () => {
         const str1 = "a;slkdfjl;askdjfkl;asdjfkl;asjdfl;";
-        const str2 = "a;sldkfjal;skdfjl;kasjdfl;k;laksjdf;laksjdf;laksjdf;lkajsd;lkfjaskl;dfjals;kdfjnal;skdjbfl;kasbdjfv;laskjbdfal;skdjfalv;skdjfalskdbjnfav;bslkdfjnalv;ksdfjbalskdfbjalvsk.dfjbalsv;kdbfjalsv;kdfjbadklsfjals";
+        const str2 =
+            "a;sldkfjal;skdfjl;kasjdfl;k;laksjdf;laksjdf;laksjdf;lkajsd;lkfjaskl;dfjals;kdfjnal;skdjbfl;kasbdjfv;laskjbdfal;skdjfalv;skdjfalskdbjnfav;bslkdfjnalv;ksdfjbalskdfbjalvsk.dfjbalsv;kdbfjalsv;kdfjbadklsfjals";
 
-        expect(utility.hashString(str1)).not.toBe(utility.hashString(str2));
+        expect(hashString(str1)).not.toBe(hashString(str2));
     });
 });
 
@@ -85,15 +86,81 @@ test("isInstanceOfOneOf", () => {
     class Thing3 {}
     class Thing4 {}
 
-    const classes = [
-        Thing1,
-        Thing2,
-        Thing3,
-        Thing4,
-    ];
+    const classes = [Thing1, Thing2, Thing3, Thing4];
 
     const thing2 = new Thing4();
 
-    expect(utility.isInstanceOfOneOf(thing2, classes)).toBe(true);
-    expect(utility.isInstanceOfOneOf(5, classes)).not.toBe(true);
+    expect(isInstanceOfOneOf(thing2, classes)).toBe(true);
+    expect(isInstanceOfOneOf(5, classes)).not.toBe(true);
+});
+
+function testSubjectsAndMatches(subjectsAndMatches: object) {
+    Object.entries(subjectsAndMatches).map(([subject, match]) => {
+        test(subject, () => {
+            const result = matchAtMention(subject, true);
+
+            if (result === null) {
+                expect(result).toBe(match);
+            } else {
+                expect(result.match).toBe(match);
+            }
+        });
+    });
+}
+
+describe("matching @mentions", () => {
+    describe("simple mentions", () => {
+        const goodSubjects = {
+            "@System": "System",
+            "Sometext @System": "System",
+            "asdfasdf @joe": "joe",
+        };
+
+        testSubjectsAndMatches(goodSubjects);
+    });
+
+    describe("special characters", () => {
+        const goodSubjects = {
+            [`@"Séche"`]: "Séche",
+            [`Something @"Séche"`]: "Séche",
+            [`@"Umuüûū"`]: "Umuüûū",
+            [`@Séche`]: "Séche", // Unquoted accent character
+            [`@Umuüûū"`]: 'Umuüûū"',
+        };
+
+        testSubjectsAndMatches(goodSubjects);
+    });
+
+    describe("names with spaces", () => {
+        const goodSubjects = {
+            [`@"Someon asdf `]: "Someon asdf ",
+            [`@"someone with a closed space"`]: "someone with a closed space",
+            [`@"What about multiple spaces?      `]: "What about multiple spaces?      ",
+        };
+
+        const badSubjects = {
+            "@someone with non-wrapped spaces": null,
+            "@Some ": null,
+        };
+
+        testSubjectsAndMatches(goodSubjects);
+        testSubjectsAndMatches(badSubjects);
+    });
+
+    describe("Closing characters", () => {
+        const goodSubjects = {
+            [`@Other Mention at end after linebreak
+                @System`]: "System",
+            [`
+    Newline with special char
+                               @"Umuüûū"`]: "Umuüûū",
+        };
+
+        const badSubjects = {
+            [`@"Close on quote" other thing`]: null,
+        };
+
+        testSubjectsAndMatches(goodSubjects);
+        testSubjectsAndMatches(badSubjects);
+    });
 });
